Rename `tags/tags` to `tags/validation`
